+++ sdk/core/Azure.Core/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Key Bug Fixes
 
 - Fixed `NotSupportedException` when running using Blazor in the browser.
+- Disable the response caching and enable the streaming when running using Blazor in the browser.
 
 ## 1.12.0 (2021-04-06)